correction of remarks

+++ controllers/cards.js
@@ -1,9 +1,8 @@
-/* eslint-disable no-underscore-dangle */
 const Card = require('../models/card');
 const Helper = require('../helper');
 
 // Получить все карточки
-module.exports.getCards = (req, res) => {
+const getCards = (req, res) => {
   Card.find({})
     .populate('owner')
     .then((cards) => res.send({ data: cards }))
@@ -11,7 +10,7 @@ module.exports.getCards = (req, res) => {
 };
 
 // Новая карточка
-module.exports.createCard = (req, res) => {
+const createCard = (req, res) => {
   const objCard = {
     name: req.body.name,
     link: req.body.link,
@@ -19,33 +18,39 @@ module.exports.createCard = (req, res) => {
   };
 
   Card.create(objCard)
-    .then((card) => res.send({ data: card }))
+    .then((card) => res.status(201).res.send({ data: card }))
     .catch((err) => res.status(Helper.getErrorNumber(err)).send({ message: err.message }));
 };
 
 // Удалить карточку
-module.exports.deleteCard = (req, res) => {
+const deleteCard = (req, res) => {
   // Находим карточку, проверяем авторизирован ее хозяин или нет
   Card.findById(req.params.id)
     .populate('owner')
     .then((card) => {
-      if (card.owner.id !== req.user._id) {
-        // Авторизирован не хозяин карточки
-        res.status(400).send({ message: 'Ошибка доступа' });
+      if (card) {
+        // Если карточка существует
+        if (card.owner.id !== req.user._id) {
+          // Авторизирован не хозяин карточки
+          res.status(403).send({ message: 'Ошибка доступа' });
+        } else {
+          // Авторизирован хозяин, удаляем карточку
+          Card.findByIdAndRemove(req.params.id)
+            .then((cardData) => res.send({ data: cardData }));
+        }
       } else {
-        // Авторизирован хозяин, удаляем карточку
-        Card.findByIdAndRemove(req.params.id)
-          .then((cardData) => res.send({ data: cardData }));
+        // Если карточка не существует
+        res.status(400).send({ message: 'Карточка с таким id не найдена' });
       }
     })
     .catch((err) => res.status(500).send({ message: err.message }));
 };
 
 // Лайкнуть карточку
-module.exports.likeCard = (req, res) => {
+const likeCard = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, {
     $addToSet: {
-      likes: req.user.id,
+      likes: req.user._id,
     },
   }, // добавить id в массив, если его там нет
   { new: true })
@@ -54,13 +59,17 @@ module.exports.likeCard = (req, res) => {
 };
 
 // Снять лайк карточке
-module.exports.dislikeCard = (req, res) => {
+const dislikeCard = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, {
     $pull: {
-      likes: req.user.id,
+      likes: req.user._id,
     },
   }, // убрать id из массива
   { new: true })
     .then((card) => res.send({ data: card }))
     .catch((err) => res.status(500).send({ message: err.message }));
+};
+
+module.exports = {
+  getCards, createCard, deleteCard, likeCard, dislikeCard,
 };
